Move opts to its own file
